Refactor `PaginatedGrid` imports across components to use `filterable-paginated-grid` module
